Delete README.md
Front End

HTML
The HyperText Markup Language, or HTML is the standard markup language for documents designed to be displayed in a web browser. 

CSS
Cascading Style Sheets is a style sheet language used for describing the presentation of a document written in a markup language such as HTML. 

BootStrap
Bootstrap is a free and open-source CSS framework directed at responsive, mobile-first front-end web development.